requirements.txt
streamlit
snowflake-connector-python[pandas] 
pandas>=2.1,<2.3
numpy>=1.26,<3
scikit-learn>=1.4,<1.6
python-dotenv>=1.0,<2
pyarrow==18.1.0                       
azure-storage-blob>=12.18
kaggle>=1.6
requests>=2.31